TechnologyReviewController and tests
